feat: chef parser
When a chef cookbook depends on an external resource, like Supermarket, we need to get that resource and list all dependencies to, at minimum, get the full picture and write much better results.

This PR checks the dependencies for the cookbook if it's part of Policyfile.rb, and reports the paths to the AI to be able to check all necessary files.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>