fix: clean warning message
When migrate, there was a Warning Message that was a bit annoying. I introduced the problem on d80842aa35c84678bd7464379cc7420e17301188 when I return the state instead of a new dictionary, the doc said this:

https://langchain-ai.github.io/langgraph/reference/agents/#langgraph.prebuilt.chat_agent_executor.create_react_agent

```
An optional node to add before the agent node (i.e., the node that calls the LLM). Useful for managing long message histories (e.g., message trimming, summarization, etc.). Pre-model hook must be a callable or a runnable that takes in current graph state and returns a state update in the form of

{
    # If provided, will UPDATE the `messages` in the state
    "messages": [RemoveMessage(id=REMOVE_ALL_MESSAGES), ...],
    # If provided, will be used as the input to the LLM,
    # and will NOT UPDATE `messages` in the state
    "llm_input_messages": [...],
    # Any other state keys that need to be propagated
    ...
}
```

With this change, the warning message is not longer there, and it's easier to debug

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>