feat: add Ansible Molecule support
When the validate_migration is triggered, and lint/check/files created
are all correct, we need to trigger the deployment to validate if it
works. This PR introduce the ansible molecule test, with podman and
different phases.

Not yet added  into src/exporters/chef_to_ansible.py because I'm waiting
for the changes in #51 to be in place, but the code it's already
working.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>