feat: Adjust logging to namespace
Currently the logs are a bit messy, so I changed all to a namespace `x2convertor.__name__`, so we can enable our logs or not based on our needs, added a new way DEBUG_ALL where enable all libraries logs levels as we need.

To avoid any issue, I also add the ruff check for checking that there is no imports from the root logging.

So, with this change:

All our log is under `x2convertor.` namespace, which are enabled by default to INFO, other logs are by default to WARNING.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>