feat: analyze always with name
When the cookbook is on the root of a project, the output for the migration plan was something like this:

`migration-plan-..md`

This change adds a way to have a proper name by the AI, and if there is no name it'll pick the path as it was. Cleaned a bit the code using the method directly into the state.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>